+++ lib/event_bus/services/store.ex
@@ -7,26 +7,26 @@ defmodule EventBus.Service.Store do
   @prefix "eb_es_"
 
   @doc false
-  @spec register_topic(String.t() | atom()) :: no_return()
-  def register_topic(topic) do
+  @spec exist?(atom()) :: boolean()
+  def exist?(topic) do
     table_name = table_name(topic)
     all_tables = :ets.all()
+    Enum.any?(all_tables, fn table -> table == table_name end)
+  end
 
-    unless Enum.any?(all_tables, fn table -> table == table_name end) do
+  @doc false
+  @spec register_topic(String.t() | atom()) :: no_return()
+  def register_topic(topic) do
+    unless exist?(topic) do
       opts = [:set, :public, :named_table, {:read_concurrency, true}]
-      Ets.new(table_name, opts)
+      Ets.new(table_name(topic), opts)
     end
   end
 
   @doc false
   @spec unregister_topic(String.t() | atom()) :: no_return()
   def unregister_topic(topic) do
-    table_name = table_name(topic)
-    all_tables = :ets.all()
-
-    if Enum.any?(all_tables, fn table -> table == table_name end) do
-      Ets.delete(table_name)
-    end
+    if exist?(topic), do: Ets.delete(table_name(topic))
   end
 
   @doc false
@@ -51,6 +51,7 @@ defmodule EventBus.Service.Store do
     :ok
   end
 
-  defp table_name(topic),
-    do: :"#{@prefix}#{topic}"
+  defp table_name(topic) do
+    :"#{@prefix}#{topic}"
+  end
 end