+++ lib/event_bus/store.ex
@@ -9,48 +9,62 @@ defmodule EventBus.Store do
 
   use GenServer
   alias EventBus.Model.Event
+  alias EventBus.Service.Store, as: StoreService
 
-  @backend Application.get_env(
-             :event_bus,
-             :store_backend,
-             EventBus.Service.Store
-           )
+  @app :event_bus
+  @backend Application.get_env(@app, :store_backend, StoreService)
 
   @doc false
-  def start_link,
-    do: GenServer.start_link(__MODULE__, nil, name: __MODULE__)
+  def start_link do
+    GenServer.start_link(__MODULE__, nil, name: __MODULE__)
+  end
 
   @doc false
-  def init(args),
-    do: {:ok, args}
+  def init(args) do
+    {:ok, args}
+  end
+
+  @doc """
+  Check if the topic exists?
+  It's important to keep this in blocking manner to prevent double creations in
+  sub modules
+  """
+  @spec exist?(String.t() | atom()) :: boolean()
+  def exist?(topic) do
+    GenServer.call(__MODULE__, {:exist?, topic})
+  end
 
   @doc """
   Register a topic to the store
   """
-  @spec register_topic(String.t() | atom()) :: no_return()
-  def register_topic(topic),
-    do: GenServer.cast(__MODULE__, {:register_topic, topic})
+  @spec register_topic(String.t() | atom()) :: :ok
+  def register_topic(topic) do
+    GenServer.call(__MODULE__, {:register_topic, topic})
+  end
 
   @doc """
   Unregister the topic from the store
   """
-  @spec unregister_topic(String.t() | atom()) :: no_return()
-  def unregister_topic(topic),
-    do: GenServer.cast(__MODULE__, {:unregister_topic, topic})
-
-  @doc """
-  Delete an event from the store
-  """
-  @spec delete({atom(), String.t() | integer()}) :: no_return()
-  def delete({topic, id}),
-    do: GenServer.cast(__MODULE__, {:delete, {topic, id}})
+  @spec unregister_topic(String.t() | atom()) :: :ok
+  def unregister_topic(topic) do
+    GenServer.call(__MODULE__, {:unregister_topic, topic})
+  end
 
   @doc """
   Save an event to the store
   """
   @spec save(Event.t()) :: :ok
-  def save(%Event{} = event),
-    do: GenServer.call(__MODULE__, {:save, event})
+  def save(%Event{} = event) do
+    GenServer.call(__MODULE__, {:save, event})
+  end
+
+  @doc """
+  Delete an event from the store
+  """
+  @spec delete({atom(), String.t() | integer()}) :: no_return()
+  def delete({topic, id}) do
+    GenServer.cast(__MODULE__, {:delete, {topic, id}})
+  end
 
   ###########################################################################
   # DELEGATIONS
@@ -69,30 +83,38 @@ defmodule EventBus.Store do
   ###########################################################################
 
   @doc false
-  @spec handle_cast({:register_topic, String.t() | atom()}, nil) :: no_return()
-  def handle_cast({:register_topic, topic}, state) do
+  @spec handle_call({:register_topic, atom()}, any(), term())
+    :: {:reply, :ok, term()}
+  def handle_call({:register_topic, topic}, _from, state) do
     @backend.register_topic(topic)
-    {:noreply, state}
+    {:reply, :ok, state}
   end
 
-  @spec handle_cast({:unregister_topic, String.t() | atom()}, nil)
-    :: no_return()
-  def handle_cast({:unregister_topic, topic}, state) do
+  @spec handle_call({:unregister_topic, atom()}, any(), term())
+    :: {:reply, :ok, term()}
+  def handle_call({:unregister_topic, topic}, _from, state) do
     @backend.unregister_topic(topic)
-    {:noreply, state}
+    {:reply, :ok, state}
   end
 
-  @spec handle_cast({:delete, {atom(), String.t() | integer()}}, nil)
+  @doc false
+  @spec handle_call({:exist?, atom()}, any(), term())
+    :: {:reply, boolean(), nil}
+  def handle_call({:exist?, topic}, _from, state) do
+    {:reply, @backend.exist?(topic), state}
+  end
+
+  @doc false
+  @spec handle_call({:save, Event.t()}, any(), term()) :: no_return()
+  def handle_call({:save, event}, _from, state) do
+    @backend.save(event)
+    {:reply, :ok, state}
+  end
+
+  @spec handle_cast({:delete, {atom(), String.t() | integer()}}, term())
     :: no_return()
   def handle_cast({:delete, {topic, id}}, state) do
     @backend.delete({topic, id})
     {:noreply, state}
-  end
-
-  @doc false
-  @spec handle_call({:save, Event.t()}, any(), nil) :: no_return()
-  def handle_call({:save, event}, _from, state) do
-    @backend.save(event)
-    {:reply, :ok, state}
   end
 end